perf(core): extract NoiseSocket and handshake runner to wacore, replace md5 crate
Extract NoiseSocket, socket error types, and standalone handshake runners
(run_xx_handshake, run_ik_handshake, select_pattern) from whatsapp-rust into
wacore (wacore::socket and wacore::handshake::runner). This allows light clients
such as wa-auth to perform Noise handshakes and framed messaging without pulling
heavy dependencies or reimplementing socket framing.

whatsapp-rust re-exports all socket and runner types with backward compatibility.
NoiseSocket constructors decouple frame tapping through the FrameTap trait,
with SentFrameTap implementing FrameTap in whatsapp-rust.

In addition, replace the external md5 crate dependency in wacore with a compact
in-tree RFC 1321 MD5 implementation, eliminating 11 KB of unoptimized symbol
bloat and dropping 9 KB from shipped wasm binaries.

Measurements:
- whatsapp-rust native stripped: 11,019,552 -> 11,010,112 bytes (-9,440 B)
- whatsapp-rust native .text: 8,872,118 -> 8,862,134 bytes (-9,984 B)
- wa-auth raw wasm: 1,102,057 -> 1,092,869 bytes (-9,188 B)
- wa-auth shipped wasm: 549,905 -> 540,818 bytes (-9,087 B)
- Noise cipher benchmarks: 0 regression
