+++ src/socket/noise_socket.rs
@@ -1,675 +1,12 @@
-use crate::socket::error::{EncryptSendError, EncryptSendErrorKind, Result, SocketError};
-use crate::transport::Transport;
-use async_channel;
-use bytes::BytesMut;
-use futures::channel::oneshot;
-use std::sync::Arc;
-use std::sync::atomic::{AtomicU32, Ordering};
-use wacore::handshake::{NoiseCipher, NoiseError};
-use wacore::libsignal::crypto::GcmInPlaceBuffer;
-use wacore::runtime::{AbortHandle, Runtime};
-
-const INLINE_ENCRYPT_THRESHOLD: usize = 16 * 1024;
-
-/// AES-GCM tag length. A frame's wire size is a fixed function of its plaintext
-/// length, which is what lets the length prefix be written before the ciphertext
-/// exists.
-const TAG_LEN: usize = 16;
-
-/// The region of the batch buffer one frame's ciphertext occupies, exposed to
-/// AES-GCM as if it were a buffer of its own.
-///
-/// Sealing through this view puts the ciphertext and its tag straight where the
-/// transport will read them. The alternative, sealing into scratch space and
-/// copying the result in, costs a second full pass over every byte sent, which
-/// is the copy comparable stacks are built to avoid: quinn seals with
-/// `PacketKey::encrypt(&self, packet, buf, header_len)` directly in the datagram
-/// buffer, and rustls encrypts each fragment into the record it will send.
-struct FrameBody<'a> {
-    out: &'a mut BytesMut,
-    /// Offset in `out` where this frame's ciphertext starts, i.e. just past its
-    /// length prefix. Held as an offset rather than a slice so the AEAD can grow
-    /// the buffer by the tag through the same view.
-    base: usize,
-}
-
-impl GcmInPlaceBuffer for FrameBody<'_> {
-    fn as_mut_slice(&mut self) -> &mut [u8] {
-        &mut self.out[self.base..]
-    }
-
-    fn as_slice(&self) -> &[u8] {
-        &self.out[self.base..]
-    }
-
-    fn resize(&mut self, new_len: usize, value: u8) {
-        self.out.resize(self.base + new_len, value);
-    }
-
-    fn truncate(&mut self, len: usize) {
-        self.out.truncate(self.base + len);
-    }
-}
-
-/// Ceilings on one batched write. They bound how much is buffered before the
-/// first frame reaches the socket; the batch never waits for work, so these
-/// only matter when a burst is already queued.
-const MAX_BATCH_FRAMES: usize = 16;
-const MAX_BATCH_WIRE_BYTES: usize = 64 * 1024;
-
-/// What the batch buffer holds between bursts: a few small stanzas coalesced.
-///
-/// One large stanza grows the buffer to its own size, and the allocation then
-/// outlives the burst that needed it, so a single media-sized send costs the
-/// session tens of KiB for the rest of the connection. Measured at 60 KiB
-/// resident per socket after one 60 KiB frame, against 8 KiB for a socket that
-/// only ever sends small ones.
-///
-/// Login alone gets there: an 812-key pre-key upload marshals to 40 799 wire
-/// bytes, ten times this capacity, before the session has sent a single
-/// message.
-const OUT_BUF_IDLE_CAPACITY: usize = 4096;
-
-/// Consecutive small batches that mark a burst as finished. Only then is the
-/// grown buffer released, so a burst spread over several batches is never
-/// interrupted to reallocate mid-flight.
-const SMALL_BATCHES_BEFORE_SHRINK: usize = 32;
-
-/// Whether the batch buffer should be swapped for an idle-sized one, advancing
-/// the burst-tracking state.
-///
-/// A free function because the buffer lives inside the sender task, where no
-/// test can reach it: the decision is only checkable if it is separable from
-/// the loop that acts on it.
-///
-/// Two conditions end a burst, and they cover opposite shapes of traffic.
-/// `queue_drained` is the sender with nothing left to write, about to block in
-/// `recv`; it is the only signal a session that grows the buffer at login and
-/// then falls silent ever emits, since it produces no further batch for a
-/// countdown over batches to advance on. The countdown covers the other shape:
-/// traffic continuous enough that the queue is never observed empty, but whose
-/// batches have shrunk back to stanza-sized.
-///
-/// `buffer_capacity` is read as well as the wire length because a frame that
-/// fails mid-encrypt is truncated back out of the buffer, leaving the
-/// allocation it grew with no wire bytes to notice it by. Only large traffic
-/// counts as the burst still running, though — a merely large buffer must not
-/// keep resetting the countdown that releases it.
-fn should_release_batch_buffer(
-    batch_wire_len: usize,
-    buffer_capacity: usize,
-    queue_drained: bool,
-    grown: &mut bool,
-    small_batches: &mut usize,
-) -> bool {
-    if batch_wire_len > OUT_BUF_IDLE_CAPACITY || buffer_capacity > OUT_BUF_IDLE_CAPACITY {
-        *grown = true;
-    }
-    // Checked ahead of the size of the batch just written: one large frame
-    // followed by silence is precisely the case to release on, and testing the
-    // size first would send it down the "burst still running" path forever.
-    if *grown && queue_drained {
-        *grown = false;
-        *small_batches = 0;
-        return true;
-    }
-    if batch_wire_len > OUT_BUF_IDLE_CAPACITY {
-        *small_batches = 0;
-        return false;
-    }
-    // An empty batch wrote nothing, so it is not evidence of anything.
-    if !*grown || batch_wire_len == 0 {
-        return false;
-    }
-    *small_batches += 1;
-    if *small_batches < SMALL_BATCHES_BEFORE_SHRINK {
-        return false;
-    }
-    *grown = false;
-    *small_batches = 0;
-    true
-}
-
-/// Result type for send operations.
-type SendResult = std::result::Result<(), EncryptSendError>;
-
-/// Wire size a plaintext will occupy once encrypted and framed: the AES-GCM tag
-/// plus the length prefix. Used to test a queued frame against the batch ceiling
-/// before paying to encrypt it.
-fn frame_wire_len(plaintext_len: usize) -> usize {
-    plaintext_len + TAG_LEN + wacore::framing::FRAME_LENGTH_SIZE
-}
-
-/// One batched write's failure, handed to every waiter in that batch.
-///
-/// `anyhow::Error` is not `Clone`, so a shared reference is what lets all the
-/// callers see the real cause instead of a re-worded copy.
-#[derive(Debug)]
-struct SharedSendFailure(Arc<EncryptSendError>);
-
-impl std::fmt::Display for SharedSendFailure {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl std::error::Error for SharedSendFailure {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        std::error::Error::source(self.0.as_ref())
-    }
-}
-
-/// A job sent to the dedicated sender task.
-struct SendJob {
-    plaintext: bytes::Bytes,
-    response_tx: oneshot::Sender<SendResult>,
-}
-
-/// What a socket reports its sends to. Both halves belong to the `Client`; a
-/// VoIP relay socket and most tests pass [`Default`], reporting to neither.
-///
-/// One struct rather than one parameter each: the observation point is shared,
-/// so the next thing that wants to watch sends plugs in here instead of widening
-/// every constructor between here and `connect()` again.
-#[derive(Default, Clone)]
-pub struct SendObservers {
-    /// Wire-byte accounting, recorded after the transport write.
-    stats: Option<Arc<wacore::stats::SessionStats>>,
-    /// Publisher for the plaintext of each frame that reached the transport.
-    sent_frames: Option<Arc<crate::client::SentFrameTap>>,
-}
-
-impl SendObservers {
-    /// Report wire bytes into `stats` and nothing else.
-    pub fn with_stats(stats: Arc<wacore::stats::SessionStats>) -> Self {
-        Self {
-            stats: Some(stats),
-            sent_frames: None,
-        }
-    }
-
-    /// Also publish each sent frame's plaintext through `tap`.
-    pub(crate) fn with_sent_frames(mut self, tap: Arc<crate::client::SentFrameTap>) -> Self {
-        self.sent_frames = Some(tap);
-        self
-    }
-}
-
-pub struct NoiseSocket {
-    read_key: Arc<NoiseCipher>,
-    read_counter: Arc<AtomicU32>,
-    /// Channel to send jobs to the dedicated sender task.
-    /// Using a channel instead of a mutex avoids blocking callers while
-    /// the current send is in progress - they can enqueue their work and
-    /// await the result without holding a lock.
-    send_job_tx: async_channel::Sender<SendJob>,
-    /// Handle to the sender task. Aborted on drop to prevent resource leaks
-    /// if the task is stuck on a slow/hanging network operation, and
-    /// explicitly by [`NoiseSocket::abort_sender`] when teardown cannot wait
-    /// for the last `Arc` to go.
-    sender_task_handle: AbortHandle,
-}
-
-impl NoiseSocket {
-    pub fn new(
-        runtime: Arc<dyn Runtime>,
-        transport: Arc<dyn Transport>,
-        write_key: NoiseCipher,
-        read_key: NoiseCipher,
-    ) -> Self {
-        Self::with_observers(
-            runtime,
-            transport,
-            write_key,
-            read_key,
-            SendObservers::default(),
-        )
-    }
-
-    /// Like [`Self::new`], reporting each send to `observers` (the main WA
-    /// session socket passes the client's; VoIP relay sockets and most tests
-    /// report to nothing).
-    pub fn with_observers(
-        runtime: Arc<dyn Runtime>,
-        transport: Arc<dyn Transport>,
-        write_key: NoiseCipher,
-        read_key: NoiseCipher,
-        observers: SendObservers,
-    ) -> Self {
-        let write_key = Arc::new(write_key);
-        let read_key = Arc::new(read_key);
-
-        // Small buffer matched to typical steady-state throughput; the sender
-        // task is network-bound (awaits `transport.send`), so a transient
-        // WebSocket stall will backpressure producers here rather than queue.
-        let (send_job_tx, send_job_rx) = async_channel::bounded::<SendJob>(8);
-
-        // Spawn the dedicated sender task
-        let transport_clone = transport.clone();
-        let write_key_clone = write_key.clone();
-        let rt_clone = runtime.clone();
-        let sender_task_handle = runtime.spawn(Box::pin(Self::sender_task(
-            rt_clone,
-            transport_clone,
-            write_key_clone,
-            send_job_rx,
-            observers,
-        )));
-
-        Self {
-            read_key,
-            read_counter: Arc::new(AtomicU32::new(0)),
-            send_job_tx,
-            sender_task_handle,
-        }
-    }
-
-    /// Dedicated sender task that processes send jobs sequentially.
-    /// This ensures frames are sent in counter order without requiring a mutex.
-    /// The task owns the write counter and processes jobs one at a time.
-    async fn sender_task(
-        runtime: Arc<dyn Runtime>,
-        transport: Arc<dyn Transport>,
-        write_key: Arc<NoiseCipher>,
-        send_job_rx: async_channel::Receiver<SendJob>,
-        observers: SendObservers,
-    ) {
-        let SendObservers { stats, sent_frames } = observers;
-        let mut write_counter: u32 = 0;
-        // BytesMut: split().freeze() yields a zero-copy Bytes while retaining
-        // the underlying allocation for the next frame.
-        let mut out_buf = BytesMut::with_capacity(OUT_BUF_IDLE_CAPACITY);
-        // Whether `out_buf` is still holding an allocation a burst grew, and how
-        // many small batches have gone out since.
-        let mut out_buf_grown = false;
-        let mut small_batches: usize = 0;
-        // A failed transport write says nothing about how much of the frame the
-        // peer received, so the counter that frame consumed can neither be
-        // reused (nonce reuse under the same write key) nor confidently skipped
-        // (the peer's read counter would desync). Both outcomes are unrecoverable
-        // in-band, so the whole sender goes out of service and the connection
-        // must be re-established with a fresh handshake key.
-        let mut poisoned = false;
-        // Reused across batches: one allocation for the life of the connection
-        // instead of one per batch.
-        let mut waiters: Vec<(oneshot::Sender<SendResult>, usize)> = Vec::new();
-        // A job pulled off the channel that would have overflowed the byte
-        // ceiling, held over to open the next batch. Dropping it (on shutdown)
-        // drops its response channel, which the caller sees as a closed sender:
-        // a held-over job can be lost, but it can never hang its caller.
-        let mut carry_over: Option<SendJob> = None;
-        // Plaintexts of this batch's frames, held only while a consumer is
-        // watching: each entry is a refcount bump on the buffer the caller
-        // marshalled, and the whole `Vec` stays empty (unallocated) otherwise.
-        // They are kept until after the write so what is published is what the
-        // transport actually accepted.
-        let mut observed: Vec<bytes::Bytes> = Vec::new();
-
-        loop {
-            let job = match carry_over.take() {
-                Some(job) => job,
-                None => match send_job_rx.recv().await {
-                    Ok(job) => job,
-                    Err(_) => break,
-                },
-            };
-            if poisoned {
-                let _ = job.response_tx.send(Err(EncryptSendError::poisoned()));
-                continue;
-            }
-
-            // Encrypt everything already queued into one buffer and write it
-            // once. Three independent producers answer a single inbound message
-            // (the reply, the delivery receipt and the stanza ack), so a write
-            // per frame turned into a syscall, a TLS record and a WebSocket
-            // message per frame. Only frames that are ALREADY waiting are taken:
-            // never block for more, or this trades syscalls for latency.
-            waiters.clear();
-            let mut encrypt_failure: Option<(oneshot::Sender<SendResult>, EncryptSendError)> = None;
-            let mut job = job;
-            loop {
-                let response_tx = job.response_tx;
-                // Cloned before the plaintext is consumed, dropped again if the
-                // frame never makes it into the buffer.
-                let to_observe = match sent_frames.as_deref() {
-                    Some(tap) if tap.enabled() => Some(job.plaintext.clone()),
-                    _ => None,
-                };
-                match Self::encrypt_frame_into(
-                    &runtime,
-                    &write_key,
-                    &mut write_counter,
-                    job.plaintext,
-                    &mut out_buf,
-                )
-                .await
-                {
-                    Ok(wire_bytes) => {
-                        waiters.push((response_tx, wire_bytes));
-                        if let Some(plaintext) = to_observe {
-                            observed.push(plaintext);
-                        }
-                    }
-                    Err(e) => {
-                        // The counter is untouched on this frame, and every
-                        // frame already in the buffer must still go out so the
-                        // peer's counters stay contiguous.
-                        encrypt_failure = Some((response_tx, e));
-                        break;
-                    }
-                }
-                if out_buf.len() >= MAX_BATCH_WIRE_BYTES || waiters.len() >= MAX_BATCH_FRAMES {
-                    break;
-                }
-                match send_job_rx.try_recv() {
-                    Ok(next) => {
-                        // Check the ceiling before appending, not after, or a
-                        // nearly-full batch overshoots it by a whole frame. A
-                        // frame that cannot fit any batch still goes alone
-                        // rather than deadlocking against the ceiling.
-                        let projected = out_buf.len() + frame_wire_len(next.plaintext.len());
-                        if projected > MAX_BATCH_WIRE_BYTES {
-                            carry_over = Some(next);
-                            break;
-                        }
-                        job = next;
-                    }
-                    Err(_) => break,
-                }
-            }
-
-            let mut batch_wire_len = 0usize;
-            let outcome = if out_buf.is_empty() {
-                Ok(())
-            } else {
-                // Zero-copy: split() hands the written bytes over and out_buf
-                // keeps its capacity for the next batch.
-                let wire = out_buf.split().freeze();
-                batch_wire_len = wire.len();
-                if waiters.len() > 1 {
-                    // The only externally visible sign that a batch happened.
-                    // Without it, "does the peer accept several frames in one
-                    // WebSocket message?" cannot be answered from a live run.
-                    log::debug!(
-                        "noise: coalesced {} frames into one {}-byte write",
-                        waiters.len(),
-                        wire.len()
-                    );
-                }
-                match transport.send(wire).await {
-                    Ok(()) => {
-                        if let Some(stats) = stats.as_deref() {
-                            for (_, wire_bytes) in &waiters {
-                                stats.record_frame_sent(*wire_bytes);
-                            }
-                        }
-                        // Re-read the gate rather than trusting the read at
-                        // capture time, so a batch that outlived its last lease
-                        // stays quiet. A release racing this instant may still
-                        // lose: the lease gates, it does not fence.
-                        if let Some(tap) = sent_frames.as_deref()
-                            && tap.enabled()
-                        {
-                            for plaintext in observed.drain(..) {
-                                tap.publish(plaintext);
-                            }
-                        }
-                        Ok(())
-                    }
-                    Err(e) => Err(EncryptSendError::transport(e)),
-                }
-            };
-            // A write that failed says nothing about what the peer received, so
-            // its frames are not reported as sent.
-            observed.clear();
-
-            {
-                // Crypto and framing failures are rejected before any byte
-                // reaches the wire and leave the counter untouched, so they do
-                // not compromise the keystream. Only a transport failure is
-                // ambiguous.
-                if let Err(err) = &outcome
-                    && matches!(err.kind, EncryptSendErrorKind::Transport)
-                {
-                    poisoned = true;
-                    // Poisoning only stops this half. A write can fail while
-                    // the read half stays open (half-open socket, or a
-                    // Transport that reports Err without emitting
-                    // Disconnected), and then nothing else would notice: the
-                    // read loop keeps running and the client reports itself
-                    // connected while every send fails forever. Closing the
-                    // transport makes the existing disconnect path observe the
-                    // drop and reconnect with a fresh handshake key, which is
-                    // the only way this sender becomes usable again.
-                    transport.disconnect().await;
-                }
-            }
-
-            // Every frame in this batch shares the fate of the single write.
-            match outcome {
-                Ok(()) => {
-                    for (response_tx, _) in waiters.drain(..) {
-                        let _ = response_tx.send(Ok(()));
-                    }
-                }
-                // One waiter owns the failure outright. This is the overwhelmingly
-                // common case, and handing over the error untouched is what keeps
-                // `err.source.downcast_ref::<MyTransportError>()` working for a
-                // caller with its own Transport: wrapping would bury the typed
-                // error one level down for no benefit, since there is nobody to
-                // share it with.
-                Err(err) if waiters.len() == 1 => {
-                    let (response_tx, _) = waiters.drain(..).next().expect("length checked");
-                    let _ = response_tx.send(Err(err));
-                }
-                // Several waiters, and EncryptSendError is not Clone: they share
-                // one Arc. Display renders only the kind, so re-wording per waiter
-                // would hand each caller "transport error" with the cause gone;
-                // sharing keeps the whole chain reachable for a refcount bump each.
-                Err(err) => {
-                    let shared = Arc::new(err);
-                    for (response_tx, _) in waiters.drain(..) {
-                        let _ = response_tx.send(Err(EncryptSendError::transport(
-                            SharedSendFailure(shared.clone()),
-                        )));
-                    }
-                }
-            }
-            if let Some((response_tx, err)) = encrypt_failure {
-                let _ = response_tx.send(Err(err));
-            }
-
-            // Release the grown allocation once the burst is over. Nothing above
-            // reads `out_buf` across iterations — `split()` left it empty — so
-            // replacing it here cannot affect a batching decision.
-            //
-            // An empty channel with nothing carried over is the end of the
-            // burst: the loop is one statement from blocking in `recv`, so no
-            // frame is in flight and no batch is pending, and releasing here
-            // interrupts nothing. `is_empty` is an instantaneous read and a
-            // producer may enqueue immediately after it — that costs one 4 KiB
-            // allocation the next batch would regrow, which is the side this
-            // trade favours: a burst that was about to continue pays once,
-            // while every session that goes quiet stops paying at all.
-            let queue_drained = carry_over.is_none() && send_job_rx.is_empty();
-            if should_release_batch_buffer(
-                batch_wire_len,
-                out_buf.capacity(),
-                queue_drained,
-                &mut out_buf_grown,
-                &mut small_batches,
-            ) {
-                out_buf = BytesMut::with_capacity(OUT_BUF_IDLE_CAPACITY);
-            }
-        }
-    }
-
-    /// Encrypt one plaintext and append the framed result to `out_buf`,
-    /// returning its wire size. The counter is burned once the framed ciphertext
-    /// is committed to `out_buf`, whether or not the write that carries it
-    /// succeeds. Every error path leaves `out_buf` exactly as it found it, which
-    /// is the only reason leaving the counter unburned there is sound: a change
-    /// that keeps partial output must burn the counter too, or the next frame
-    /// reuses its nonce.
-    async fn encrypt_frame_into(
-        runtime: &Arc<dyn Runtime>,
-        write_key: &Arc<NoiseCipher>,
-        write_counter: &mut u32,
-        plaintext: bytes::Bytes,
-        out_buf: &mut BytesMut,
-    ) -> std::result::Result<usize, EncryptSendError> {
-        let counter = *write_counter;
-        // Refuse to wrap the per-direction frame counter: reusing an AES-GCM
-        // nonce under the same key is catastrophic. 2^32 frames per connection
-        // is unreachable in practice, so erroring here forces a reconnect
-        // rather than a silent nonce reuse.
-        if counter == u32::MAX {
-            return Err(EncryptSendError::crypto(NoiseError::CounterExhausted));
-        }
-        let before = out_buf.len();
-
-        if plaintext.len() <= INLINE_ENCRYPT_THRESHOLD {
-            // Ciphertext is exactly the plaintext plus the tag, so the length
-            // prefix is known before the bytes it counts exist and the frame can
-            // be sealed where it already sits in the batch.
-            let body_len = plaintext.len() + TAG_LEN;
-            if let Err(e) = wacore::framing::append_frame_header_into(body_len, None, out_buf) {
-                return Err(EncryptSendError::framing(e));
-            }
-            let base = out_buf.len();
-            out_buf.extend_from_slice(&plaintext);
-            if let Err(e) = write_key
-                .encrypt_in_place_with_counter(counter, &mut FrameBody { out: out_buf, base })
-            {
-                // Unlike the paths above, this one has already appended the
-                // prefix and the plaintext. Rolling both back is what keeps the
-                // rest of the batch, which still has to go out, contiguous, and
-                // what keeps this counter safe to hand to the next frame. The
-                // default AEAD cannot fail on a fixed-size key and nonce, so
-                // only a `set_crypto_provider` backend reaches this: it is the
-                // contract for those, not dead code.
-                out_buf.truncate(before);
-                return Err(EncryptSendError::crypto(e));
-            }
-            // The length prefix was written from `plaintext.len() + TAG_LEN`
-            // before the ciphertext existed, which is sound only because
-            // `TransportAead` is AES-256-GCM by contract. A `set_crypto_provider`
-            // backend that grows the buffer by anything else would put a frame
-            // on the wire whose prefix disagrees with its body, desyncing the
-            // peer's parser for the rest of the connection. Checking costs one
-            // comparison and turns that into a refused send.
-            if out_buf.len() - base != body_len {
-                out_buf.truncate(before);
-                return Err(EncryptSendError::crypto(NoiseError::Encrypt(
-                    wacore::libsignal::crypto::CryptoProviderError::BackendFailed,
-                )));
-            }
-        } else {
-            let write_key = write_key.clone();
-            // `Bytes` is Send + 'static: move it into the blocking task (a
-            // refcount bump) instead of copying the whole >16KB plaintext.
-            let encrypt_result = wacore::runtime::blocking(&**runtime, move || {
-                write_key.encrypt_with_counter(counter, &plaintext)
-            })
-            .await;
-            let ciphertext = match encrypt_result {
-                Ok(c) => c,
-                Err(e) => return Err(EncryptSendError::crypto(e)),
-            };
-            if let Err(e) = wacore::framing::append_frame_into(&ciphertext, None, out_buf) {
-                return Err(EncryptSendError::framing(e));
-            }
-        }
-
-        *write_counter = counter + 1;
-        Ok(out_buf.len() - before)
-    }
-
-    /// Hands `plaintext` to the sender task and returns the channel its result
-    /// will arrive on, without waiting for it.
-    ///
-    /// Split out of [`Self::encrypt_and_send`] so a burst can enqueue every
-    /// frame before awaiting any of them. The sender coalesces whatever is
-    /// already queued into one transport write, so a caller that awaited each
-    /// frame before enqueueing the next would hand them over one completion
-    /// apart and get one write per frame.
-    ///
-    /// The returned receiver must be awaited, or the result is dropped and the
-    /// caller cannot tell a delivered frame from a failed one.
-    pub(crate) async fn enqueue_send(
-        &self,
-        plaintext: bytes::Bytes,
-    ) -> std::result::Result<oneshot::Receiver<SendResult>, EncryptSendError> {
-        let (response_tx, response_rx) = oneshot::channel();
-
-        let job = SendJob {
-            plaintext,
-            response_tx,
-        };
-
-        // Send job to the sender task. If channel is closed, sender task has stopped.
-        if let Err(_send_err) = self.send_job_tx.send(job).await {
-            return Err(EncryptSendError::channel_closed());
-        }
-
-        Ok(response_rx)
-    }
-
-    /// Awaits a receiver handed out by [`Self::enqueue_send`].
-    pub(crate) async fn await_send(receiver: oneshot::Receiver<SendResult>) -> SendResult {
-        match receiver.await {
-            Ok(result) => result,
-            Err(_) => {
-                // Sender task dropped without sending a response
-                Err(EncryptSendError::channel_closed())
-            }
-        }
-    }
-
-    pub async fn encrypt_and_send(&self, plaintext: bytes::Bytes) -> SendResult {
-        let receiver = self.enqueue_send(plaintext).await?;
-        Self::await_send(receiver).await
-    }
-
-    /// Stop the sender task now, without waiting for the last `Arc` to drop.
-    ///
-    /// Drop alone is not enough on a teardown path. The socket is handed out as
-    /// an `Arc`, and a send held across an await (`send_raw_bytes`) keeps a
-    /// clone alive for as long as that send is parked — which, on a black-holed
-    /// TCP connection, is however long the kernel takes to give up on the write
-    /// (`tcp_retries2`, ~15 minutes on Linux). Until then the task stays parked
-    /// inside `transport.send`, holding the transport's sink, so the close that
-    /// teardown is waiting on cannot even start. Cancelling the task drops that
-    /// send, releases the sink, and hands every queued and in-flight caller a
-    /// closed channel instead of a wait with no deadline.
-    ///
-    /// Closing the job channel first is what makes the failure prompt rather
-    /// than merely eventual: an enqueue racing the abort fails immediately
-    /// instead of landing in a queue nothing will drain.
-    pub(crate) fn abort_sender(&self) {
-        self.send_job_tx.close();
-        self.sender_task_handle.abort();
-    }
-
-    pub fn decrypt_frame(&self, mut ciphertext: BytesMut) -> Result<BytesMut> {
-        // Checked increment: error instead of wrapping the read counter (AES-GCM
-        // nonce reuse). fetch_update returns the pre-increment counter to use, or
-        // Err when it would overflow u32. Mirrors the write side.
-        let counter = self
-            .read_counter
-            .fetch_update(Ordering::SeqCst, Ordering::SeqCst, |c| c.checked_add(1))
-            .map_err(|_| SocketError::Cipher(NoiseError::CounterExhausted))?;
-        self.read_key
-            .decrypt_in_place_with_counter(counter, &mut ciphertext)
-            .map_err(SocketError::Cipher)?;
-        Ok(ciphertext)
-    }
-}
-
-// AbortHandle aborts the sender task on drop automatically, so no manual
-// Drop impl is needed — the `sender_task_handle` field's own Drop does the work.
-// `abort_sender` is the same stop, taken early by a teardown that cannot wait
-// for the last `Arc` clone to be released.
+pub use crate::transport::Transport;
+pub use bytes::BytesMut;
+pub use std::sync::Arc;
+pub use std::sync::atomic::AtomicU32;
+pub use wacore::handshake::{NoiseCipher, NoiseError};
+pub use wacore::libsignal::crypto::GcmInPlaceBuffer;
+pub use wacore::runtime::Runtime;
+pub use wacore::socket::error::*;
+pub use wacore::socket::noise_socket::*;
 
 #[cfg(test)]
 mod tests {
