+++ wacore/Cargo.toml
@@ -72,7 +72,6 @@ hkdf = { workspace = true }
 hmac = { workspace = true }
 itoa = { workspace = true }
 log = { workspace = true }
-md5 = "0.8.1"
 metrics = { workspace = true, optional = true }
 portable-atomic = { workspace = true }
 rand = { workspace = true }
