+++ wacore/src/handshake/mod.rs
@@ -6,3 +6,9 @@ pub use wacore_noise::{
     WA_CERT_PUB_KEY, XxFallbackHandshakeState, XxHandshakeOutcome, XxHandshakeState,
     build_edge_routing_preintro, build_handshake_header, generate_iv,
 };
+
+pub mod runner;
+pub use runner::{
+    HandshakeError as HandshakeExecutionError, HandshakePattern, HandshakeSuccess, recv_frame,
+    run_ik_handshake, run_xx_handshake, select_pattern, send_first_handshake_message,
+};
